Improved directory structure & code separation
