+++ lib/app/model/navigation_tab_model.dart
@@ -1,10 +1,14 @@
 import 'package:flutter/cupertino.dart';
 
-class NavigationTab {
-  final Icon icon;
+class NavigationTabModel {
+  final Widget icon;
   final String label;
   final Widget child;
   final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
-  NavigationTab({required this.icon, required this.label, required this.child});
+  NavigationTabModel({
+    required this.icon,
+    required this.label,
+    required this.child,
+  });
 }
