+++ lib/enums/task_status.dart
@@ -1,0 +1,7 @@
+enum TaskListSection {
+  requests,
+  finished,
+  rejected,
+  failed,
+  archived,
+}
