+++ lib/pages/device_page.dart
