+++ lib/util/card_reader_launcher.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart' hide Card;
 import 'package:meesign_core/meesign_card.dart';
-import '../ui/card_reader_page.dart';
+import '../pages/card_reader_page.dart';
 
 void launchCardReader(
   BuildContext context,
