+++ lib/util/extensions/task_approvable.dart
@@ -1,0 +1,6 @@
+import 'package:meesign_core/meesign_core.dart';
+
+extension TaskDetails on Task {
+  bool get approvable =>
+      !approved && (state == TaskState.created || state == TaskState.running);
+}
